// projects/egeo/src/lib/st-breadcrumbs/st-breadcrumbs.interface.ts
/*
 * © 2017 Stratio Big Data Inc., Sucursal en España.
 *
 * This software is licensed under the Apache License, Version 2.0.
 * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
 * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
 * See the terms of the License for more details.
 *
 * SPDX-License-Identifier: Apache-2.0.
 */

export class StBreadCrumbItem {
   id: string;
   label?: string;
   icon?: string;
}

export enum StBreadCrumbMode { DEFAULT =  <any> 'default', TITLE =  <any> 'title'}